Move RestCountAction to the 'search' package.
